requirements.txt
apscheduler
holidays
pyrogram
python-dotenv
pytz
requests
tgcrypto
uvloop
